requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cheroot==8.5.2
CherryPy==18.6.0
configparser==5.0.1
Django==3.1.7
django-minio-storage==0.3.10
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
minio==6.0.2
more-itertools==8.7.0
portend==2.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tempora==4.0.1
urllib3==1.26.3
zc.lockfile==2.0
psycopg2-binary==2.9.5
Pillow
django-storages